+++ src/components/picker/picker.vue
@@ -37,6 +37,7 @@
   import BScroll from 'better-scroll'
   import CubePopup from '../popup/popup.vue'
   import apiMixin from '../../common/mixins/api'
+  import pickerMixin from '../../common/mixins/picker'
 
   const COMPONENT_NAME = 'cube-picker'
 
@@ -45,64 +46,13 @@
   const EVENT_CANCEL = 'cancel'
   const EVENT_CHANGE = 'change'
 
-  const DEFAULT_KEYS = {
-    value: 'value',
-    text: 'text'
-  }
-
   export default {
     name: COMPONENT_NAME,
-    mixins: [apiMixin],
-    props: {
-      data: {
-        type: Array,
-        default() {
-          return []
-        }
-      },
-      selectedIndex: {
-        type: Array,
-        default() {
-          return []
-        }
-      },
-      title: {
-        type: String
-      },
-      cancelTxt: {
-        type: String,
-        default: '取消'
-      },
-      confirmTxt: {
-        type: String,
-        default: '确定'
-      },
-      swipeTime: {
-        type: Number,
-        default: 2500
-      },
-      alias: {
-        type: Object,
-        default() {
-          return {}
-        }
-      },
-      zIndex: {
-        type: Number
-      }
-    },
+    mixins: [apiMixin, pickerMixin],
     data() {
       return {
         pickerData: this.data.slice(),
         pickerSelectedIndex: this.selectedIndex
-      }
-    },
-    computed: {
-      valueKey() {
-        return this.alias.value || DEFAULT_KEYS.value
-      },
-      textKey() {
-        return this.alias.text || DEFAULT_KEYS.text
       }
     },
     created() {
@@ -291,14 +241,6 @@
         return this.wheels.every((wheel) => {
           return !wheel.isInTransition
         })
-      }
-    },
-    watch: {
-      data(newVal) {
-        this.setData(newVal, this.selectedIndex)
-      },
-      selectedIndex(newVal) {
-        this.setData(this.data, newVal)
       }
     },
     components: {
